feat: Update correosController.js to include mailList in notification emails
The recent changes to correosController.js include adding a query to retrieve email addresses from the secretaria table and storing them in the resultsmail variable. The mailList variable is then created by mapping the mail property of each row in resultsmail. The cc field in the email notification is updated to include all email addresses in mailList, separated by commas.

This commit message suggests updating correosController.js to include the mailList variable in the cc field of notification emails.