feat: Add PATCH method to allowed HTTP methods in CORS options
This commit adds the PATCH method to the list of allowed HTTP methods in the CORS options configuration. Previously, only GET, PUT, POST, and DELETE methods were allowed. With this change, the application will now support the PATCH method as well.

This commit message suggests including the PATCH method in the allowed HTTP methods for CORS options.